Remove alternate words processing.
Alternate words reduced processing time by not recalculating an anagram
for the second time of two words in the anagram were anagrams of
themselves. Ex:

A BRAD DAY
A BARD DAY

In this case the second case would not be calculated, but instead
derived from the first.

The problem happens when we have a valid anagram with a word in the list
of alternates. Ex:

A BRAD BARD DAY
A BARD BRAD DAY

In this case, the second permutation would not print at all, since
"BRAD" and "BARD" are never considered together (they're anagrams of
each other).

Removing this feature increased the processing time.